+++ src/test/java/lotto/model/BonusNumberTest.java
@@ -1,7 +1,6 @@
-package lotto;
+package lotto.model;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import lotto.model.BonusNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
